+++ src/main/java/roomescape/time/dto/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.time.application;
+package roomescape.time.dto;
 
 import java.time.LocalTime;
